Threat detection
The program defines a list of known threats as an array of strings. It then prompts the user to enter a string, which it scans for any of the known threats. If a threat is found, the program prints the name of the threat and exits. This as a base to build more complex threat detection applications.